Move inline functions of Node class from ContainerNode.h to NodeInlines.h https://bugs.webkit.org/show_bug.cgi?id=289026
Reviewed by Ryosuke Niwa.

Some inline functions of Node class were defined in ContainerNode.h.
<https://commits.webkit.org/291488@main> added NodeInlines.h. Move the
inline functions to NodeInlines.h.

* Source/WebCore/SaferCPPExpectations/UncountedCallArgsCheckerExpectations:
* Source/WebCore/accessibility/AccessibilityMathMLElement.cpp:
* Source/WebCore/accessibility/AccessibilityTree.cpp:
* Source/WebCore/bindings/js/JSNodeCustom.h:
* Source/WebCore/dom/ContainerNode.h:
(WebCore::Node::countChildNodes const): Deleted.
(WebCore::Node::traverseToChildAt const): Deleted.
(WebCore::Node::firstChild const): Deleted.
(WebCore::Node::protectedFirstChild const): Deleted.
(WebCore::Node::lastChild const): Deleted.
(WebCore::Node::protectedLastChild const): Deleted.
(WebCore::Node::hasChildNodes const): Deleted.
(WebCore::Node::rootNode const): Deleted.
(WebCore::collectChildNodes): Deleted.
(WebCore::Node::setParentNode): Deleted.
(WebCore::Node::protectedParentNode const): Deleted.
* Source/WebCore/dom/Element.cpp:
* Source/WebCore/dom/ElementAndTextDescendantIterator.h:
* Source/WebCore/dom/EventPath.cpp:
* Source/WebCore/dom/Node.h:
* Source/WebCore/dom/NodeInlines.h:
(WebCore::Node::countChildNodes const):
(WebCore::Node::traverseToChildAt const):
(WebCore::Node::firstChild const):
(WebCore::Node::protectedFirstChild const):
(WebCore::Node::lastChild const):
(WebCore::Node::protectedLastChild const):
(WebCore::Node::hasChildNodes const):
(WebCore::Node::rootNode const):
(WebCore::Node::setParentNode):
(WebCore::Node::protectedParentNode const):
(WebCore::collectChildNodes):
* Source/WebCore/dom/NodeTraversal.h:
* Source/WebCore/dom/Position.h:
(WebCore::offsetIsBeforeLastNodeOffset): Deleted.
* Source/WebCore/dom/PositionInlines.h:
(WebCore::offsetIsBeforeLastNodeOffset):
* Source/WebCore/dom/PositionIterator.cpp:
(WebCore::PositionIterator::PositionIterator):
* Source/WebCore/dom/PositionIterator.h:
(WebCore::PositionIterator::PositionIterator): Deleted.
* Source/WebCore/dom/RangeBoundaryPoint.h:
(WebCore::RangeBoundaryPoint::set): Deleted.
* Source/WebCore/dom/RangeBoundaryPointInlines.h:
(WebCore::RangeBoundaryPoint::set):
* Source/WebCore/editing/ApplyStyleCommand.cpp:
* Source/WebCore/editing/RemoveNodePreservingChildrenCommand.cpp:
* Source/WebCore/editing/RenderedPosition.cpp:
* Source/WebCore/editing/ReplaceNodeWithSpanCommand.cpp:
* Source/WebCore/page/SpatialNavigation.cpp:
* Source/WebCore/xml/XPathPath.cpp:
* Source/WebKit/WebProcess/InjectedBundle/API/mac/WKDOMNode.mm:
* Tools/TestWebKitAPI/Tests/WebCore/HTMLParserIdioms.cpp:

Canonical link: https://commits.webkit.org/291613@main
