+++ Source/WebCore/dom/Position.h
@@ -322,16 +322,7 @@ inline Position firstPositionInNode(Node* anchorNode)
     return Position(anchorNode, Position::PositionIsBeforeChildren);
 }
 
-inline bool offsetIsBeforeLastNodeOffset(unsigned offset, Node* anchorNode)
-{
-    if (auto* characterData = dynamicDowncast<CharacterData>(*anchorNode))
-        return offset < characterData->length();
-
-    unsigned currentOffset = 0;
-    for (Node* node = anchorNode->firstChild(); node && currentOffset < offset; node = node->nextSibling())
-        currentOffset++;
-    return offset < currentOffset;
-}
+inline bool offsetIsBeforeLastNodeOffset(unsigned offset, Node* anchorNode);
 
 } // namespace WebCore
