+++ Source/WebCore/dom/PositionInlines.h
@@ -37,4 +37,15 @@ Position lastPositionInNode(Node* anchorNode)
     return Position(anchorNode, Position::PositionIsAfterChildren);
 }
 
+inline bool offsetIsBeforeLastNodeOffset(unsigned offset, Node* anchorNode)
+{
+    if (auto* characterData = dynamicDowncast<CharacterData>(*anchorNode))
+        return offset < characterData->length();
+
+    unsigned currentOffset = 0;
+    for (Node* node = anchorNode->firstChild(); node && currentOffset < offset; node = node->nextSibling())
+        currentOffset++;
+    return offset < currentOffset;
+}
+
 } // namespace WebCore
