+++ Source/WebCore/dom/PositionIterator.cpp
@@ -41,6 +41,13 @@ namespace WebCore {
 
 using namespace HTMLNames;
 
+PositionIterator::PositionIterator(const Position& pos)
+    : m_anchorNode(pos.anchorNode())
+    , m_nodeAfterPositionInAnchor(m_anchorNode->traverseToChildAt(pos.deprecatedEditingOffset()))
+    , m_offsetInAnchor(m_nodeAfterPositionInAnchor ? 0 : pos.deprecatedEditingOffset())
+{
+}
+
 PositionIterator::operator Position() const
 {
     auto anchorNode = protectedNode();
