+++ Source/WebCore/editing/ApplyStyleCommand.cpp
@@ -41,6 +41,7 @@
 #include "HTMLNames.h"
 #include "HTMLSpanElement.h"
 #include "LocalFrame.h"
+#include "NodeInlines.h"
 #include "NodeList.h"
 #include "NodeTraversal.h"
 #include "RenderObject.h"
